#98 [FEAT] 결제 대기 중 Loading 컴포넌트 노출
