+++ src/components/Animation/Animation.jsx
@@ -1,0 +1,14 @@
+import Lottie from 'react-lottie';
+
+export default function Animation({ animationData }) {
+  const defaultOptions = {
+    loop: true,
+    autoplay: true,
+    animationData,
+    rendererSettings: {
+      preserveAspectRatio: 'xMidYMid slice',
+    },
+  };
+
+  return <Lottie options={defaultOptions} height={400} width={400} />;
+}
