+++ src/components/Animation/index.js
@@ -1,0 +1,1 @@
+export { default as Animation } from './Animation.jsx';
